refactor: adapter -> adaptor 수정
