+++ src/main/java/com/nhnacademy/inkbridge/front/adaptor/PointPolicyTypeAdaptor.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.inkbridge.front.adapter;
+package com.nhnacademy.inkbridge.front.adaptor;
 
 import com.nhnacademy.inkbridge.front.dto.pointpolicytype.PointPolicyTypeCreateRequestDto;
 import com.nhnacademy.inkbridge.front.dto.pointpolicytype.PointPolicyTypeReadResponseDto;
@@ -10,7 +10,7 @@ import java.util.List;
  * @author jangjaehun
  * @version 2024/02/22
  */
-public interface PointPolicyTypeAdapter {
+public interface PointPolicyTypeAdaptor {
 
     void createPointPolicyType(PointPolicyTypeCreateRequestDto requestDto);
