+++ src/main/java/com/nhnacademy/inkbridge/front/service/impl/PointPolicyTypeServiceImpl.java
@@ -1,6 +1,6 @@
 package com.nhnacademy.inkbridge.front.service.impl;
 
-import com.nhnacademy.inkbridge.front.adapter.PointPolicyTypeAdapter;
+import com.nhnacademy.inkbridge.front.adaptor.PointPolicyTypeAdaptor;
 import com.nhnacademy.inkbridge.front.dto.pointpolicytype.PointPolicyTypeCreateRequestDto;
 import com.nhnacademy.inkbridge.front.dto.pointpolicytype.PointPolicyTypeReadResponseDto;
 import com.nhnacademy.inkbridge.front.service.PointPolicyTypeService;
@@ -18,7 +18,7 @@ import org.springframework.stereotype.Service;
 @RequiredArgsConstructor
 public class PointPolicyTypeServiceImpl implements PointPolicyTypeService {
 
-    private final PointPolicyTypeAdapter pointPolicyTypeAdapter;
+    private final PointPolicyTypeAdaptor pointPolicyTypeAdapter;
 
     /**
      * 포인트 정책 유형을 생성하는 메소드입니다.
